Move `physical_plan::file_format` to `datasource::plan`
