+++ datafusion/core/src/datasource/mod.rs
@@ -28,6 +28,7 @@ pub mod file_format;
 pub mod listing;
 pub mod listing_table_factory;
 pub mod memory;
+pub mod physical_plan;
 pub mod streaming;
 pub mod view;
