+++ datafusion/core/src/datasource/physical_plan/avro.rs
@@ -165,8 +165,8 @@ impl ExecutionPlan for AvroExec {
 #[cfg(feature = "avro")]
 mod private {
     use super::*;
-    use crate::physical_plan::file_format::file_stream::{FileOpenFuture, FileOpener};
-    use crate::physical_plan::file_format::FileMeta;
+    use crate::datasource::physical_plan::file_stream::{FileOpenFuture, FileOpener};
+    use crate::datasource::physical_plan::FileMeta;
     use bytes::Buf;
     use futures::StreamExt;
     use object_store::{GetResult, ObjectStore};
@@ -222,7 +222,7 @@ mod tests {
     use crate::datasource::file_format::{avro::AvroFormat, FileFormat};
     use crate::datasource::listing::PartitionedFile;
     use crate::datasource::object_store::ObjectStoreUrl;
-    use crate::physical_plan::file_format::chunked_store::ChunkedStore;
+    use crate::datasource::physical_plan::chunked_store::ChunkedStore;
     use crate::prelude::SessionContext;
     use crate::scalar::ScalarValue;
     use crate::test::object_store::local_unpartitioned_file;
