+++ datafusion/core/src/datasource/physical_plan/parquet/page_filter.rs
@@ -39,10 +39,10 @@ use parquet::{
 };
 use std::sync::Arc;
 
-use crate::physical_optimizer::pruning::{PruningPredicate, PruningStatistics};
-use crate::physical_plan::file_format::parquet::{
+use crate::datasource::physical_plan::parquet::{
     from_bytes_to_i128, parquet_to_arrow_decimal_type,
 };
+use crate::physical_optimizer::pruning::{PruningPredicate, PruningStatistics};
 
 use super::metrics::ParquetFileMetrics;
