+++ datafusion/core/src/test_util/parquet.rs
@@ -26,11 +26,11 @@ use crate::common::ToDFSchema;
 use crate::config::ConfigOptions;
 use crate::datasource::listing::{ListingTableUrl, PartitionedFile};
 use crate::datasource::object_store::ObjectStoreUrl;
+use crate::datasource::physical_plan::{FileScanConfig, ParquetExec};
 use crate::error::Result;
 use crate::optimizer::simplify_expressions::{ExprSimplifier, SimplifyContext};
 use crate::physical_expr::create_physical_expr;
 use crate::physical_expr::execution_props::ExecutionProps;
-use crate::physical_plan::file_format::{FileScanConfig, ParquetExec};
 use crate::physical_plan::filter::FilterExec;
 use crate::physical_plan::metrics::MetricsSet;
 use crate::physical_plan::ExecutionPlan;
