+++ datafusion/core/tests/parquet/schema_coercion.rs
@@ -21,8 +21,8 @@ use arrow_array::types::Int32Type;
 use arrow_array::{ArrayRef, DictionaryArray, Float32Array, Int64Array, StringArray};
 use arrow_schema::DataType;
 use datafusion::assert_batches_sorted_eq;
+use datafusion::datasource::physical_plan::{FileScanConfig, ParquetExec};
 use datafusion::physical_plan::collect;
-use datafusion::physical_plan::file_format::{FileScanConfig, ParquetExec};
 use datafusion::prelude::SessionContext;
 use datafusion_common::Result;
 use datafusion_common::Statistics;
